selecttest/unload.html

<!doctype html>
<title>Selection unload tests</title>
<div id=log></div>
<script src=http://w3c-test.org/resources/testharness.js></script>
<script>
"use strict";

// This tests the requirement "As one of the unloading document cleanup steps,
// the user agent must remove any range from the selection associated with the
// Document's browsing context (so the selection's range is null) and set its
// direction to forwards."  To do so, it creates an iframe and runs
// document.open().
//
// TODO: Test additional ways of unloading the document, like back and forward.
// TODO: Test that selection direction is reset.

var iframe = document.createElement("iframe");
var t = async_test("Selection.rangeCount must be reset to 0 after document.open()");
iframe.onload = function() {
	t.step(function() {
		assert_equals(iframe.contentWindow.getSelection().rangeCount, 0,
			"Sanity check: rangeCount is initially 0");
		iframe.contentDocument.body.appendChild(
			iframe.contentDocument.createTextNode("foo"));
		var range = iframe.contentDocument.createRange();
		range.selectNodeContents(iframe.contentDocument.body);
		iframe.contentWindow.getSelection().addRange(range);
		assert_equals(iframe.contentWindow.getSelection().rangeCount, 1,
			"Sanity check: rangeCount is 1 after adding range");
		iframe.contentDocument.open();
		assert_equals(iframe.contentWindow.getSelection().rangeCount, 0,
			"All ranges must be removed from the selection on unload");
	});
	t.done();
	document.body.removeChild(iframe);
};
document.body.appendChild(iframe);
</script>
